cast away dumb compiler complaint

+++ filters/GridDecimationFilter.cpp
@@ -78,7 +78,7 @@ void GridDecimationFilter::processOne(BOX2D bounds, PointRef& point, PointViewPt
     //get the grid cell
     double x = point.getFieldAs<double>(Dimension::Id::X);
     double y = point.getFieldAs<double>(Dimension::Id::Y);
-    int id = point.getFieldAs<double>(Dimension::Id::PointId);
+    int id = (int)point.getFieldAs<double>(Dimension::Id::PointId);
 
     double d_width_pt = std::floor((x - bounds.minx) / m_args->m_edgeLength) + 1;
     double d_height_pt = std::floor((y - bounds.miny) / m_args->m_edgeLength) + 1;
